Move app factory to __init__
This commit moves the Application Factory into AIPscan's __init__
module. This seems to be common practice, reduces the number of new
modules added in this work, and provides the potential down the line
to call the Application Factory directly from gunicorn without the need
for a wsgi.py entrypoint should we choose to do so.
